form attribute split and recent record fix
<!-- ps-id: ff71dbad-b449-48db-a3bc-7485f4428c71 -->
