--- packages/types/src/AdminFieldType.ts
@@ -1,1 +1,0 @@
-
