+++ packages/types/src/App.ts
@@ -1,0 +1,1 @@
+export type MySharedType = {};
